Move VisualStudioTaskSchedulerProvider to the EditorFeatures layer
Fixes #48644
